pyproject.toml
[tool.poetry]
name = "j2dx"
version = "2.0.0"
description = "Joy2DroidX server for virtual gamepad emulation"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
# Versioni testate che funzionano insieme
python-socketio = "5.8.0"  # Mantenuta questa versione stabile
python-engineio = "4.4.1"  # Mantenuta questa versione compatibile
uvicorn = {version = "0.22.0", extras = ["standard"]}
websockets = "11.0.3"
h11 = "0.14.0"
click = "8.1.7"
qrcode = "7.4.2"
evdev = "1.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
j2dx = "j2dx:main"

[virtualenvs]
in-project = true
create = true
